Tests for auth

+++ Address_Book/src/main/java/CSCB532/Address_Book/customRow/CustomRowService.java
@@ -30,7 +30,6 @@ public class CustomRowService {
     }
 
     public DtoCustomRow createCustomRow(DtoCustomRow dtoCustomRow) {
-        validateCustomRowInput(dtoCustomRow);
         validateContactExists(dtoCustomRow.getContactId());
         validateUserPermission(dtoCustomRow.getContactId());
 
@@ -38,12 +37,7 @@ public class CustomRowService {
                 .orElseThrow(() -> new ContactNotFoundException("Contact with ID " + dtoCustomRow.getContactId() + " not found."));
 
         ModelMapper modelMapper = new ModelMapper();
-        modelMapper.addMappings(new PropertyMap<DtoCustomRow, CustomRow>() {
-            @Override
-            protected void configure() {
-                skip(destination.getId());
-            }
-        });
+
 
         CustomRow customRow = modelMapper.map(dtoCustomRow, CustomRow.class);
         customRow.setContact(contact);
@@ -58,14 +52,44 @@ public class CustomRowService {
 
 
     public DtoCustomRow updateCustomRow(Integer id, DtoCustomRow dtoCustomRow) {
-        if (id == null || id < 0) {
-            throw new BadRequestException("Invalid custom row ID.");
-        }
 
         CustomRow existingCustomRow = customRowRepository.findById(id)
                 .orElseThrow(() -> new CustomRowNotFoundException("Custom row not found for ID: " + id));
 
-        validateUserPermission(existingCustomRow.getContact().getId());
+        boolean isBlank = false;
+        if (dtoCustomRow.getCustomField() != null){
+            if (dtoCustomRow.getCustomField().isBlank()){
+                isBlank = true;
+//                throw new BadRequestException("Field Name can't be blank.");
+            }
+            if (existingCustomRow.getCustomField().equals(dtoCustomRow.getCustomField())){
+                dtoCustomRow.setCustomField(null);
+//                throw new BadRequestException("Field Name can't be the same.");
+            }else if(isBlank){
+                dtoCustomRow.setCustomField(null);
+            }
+        }
+
+        boolean isNameBlank = false;
+        if ( dtoCustomRow.getCustomName() != null){
+            if (dtoCustomRow.getCustomName().isBlank()){
+                isNameBlank = true;
+//                throw new BadRequestException("Custom Name can't be empty.");
+            }
+            if (existingCustomRow.getCustomName().equals(dtoCustomRow.getCustomName())){
+                dtoCustomRow.setCustomName(null);
+//                    throw new BadRequestException("Custom Name can't be the same.");
+            }else if(isNameBlank){
+                dtoCustomRow.setCustomName(null);
+            }
+        }
+
+        if (dtoCustomRow.getCustomName() == null && dtoCustomRow.getCustomField() == null){
+            throw new BadRequestException("Incorrect request body");
+        }
+
+
+            validateUserPermission(existingCustomRow.getContact().getId());
 
         ModelMapper modelMapper = configureModelMapperForUpdate();
         modelMapper.map(dtoCustomRow, existingCustomRow);
@@ -120,6 +144,9 @@ public class CustomRowService {
         validateUserPermission(contactId);
 
         List<CustomRow> customRows = customRowRepository.findByContactId(contactId);
+        if (customRows.isEmpty()){
+            throw new CustomRowNotFoundException("No Custom Rows found for contact with id " + contactId);
+        }
         try {
             customRowRepository.deleteAll(customRows);
         } catch (DataAccessException exc) {
@@ -137,18 +164,7 @@ public class CustomRowService {
         }
     }
 
-    private void validateCustomRowInput(DtoCustomRow dtoCustomRow) {
-        if (dtoCustomRow.getContactId() == null) {
-            throw new BadRequestException("Missing contact ID");
-        } else if (dtoCustomRow.getContactId() < 0) {
-            throw new BadRequestException("Invalid Contact ID: is negative value.");
-        } else if (dtoCustomRow.getCustomField() == null ||
-                dtoCustomRow.getCustomName() == null ||
-                dtoCustomRow.getCustomName().trim().isEmpty() ||
-                dtoCustomRow.getCustomField().trim().isEmpty()) {
-            throw new BadRequestException("Missing input: every field must have a value.");
-        }
-    }
+
 
     private void validateContactExists(Integer contactId) {
         if (contactRepository.findById(contactId).isEmpty()) {
